Version 1.3 : basic instruction (except sqrt) handle normal float do not handle exceptions

+++ tests/matmul/matmul.c
@@ -21,7 +21,7 @@ int main(void)
     int i;
     int j;
     int k;
-    float sum;
+    int sum;
 
     for (i=0; i<4; i++)
     {
